add first refactor in sensititivy analysis
