+++ src/main/java/com/powsybl/openloadflow/sensi/OpenSensitivityAnalysisProvider.java
@@ -15,6 +15,8 @@ import com.powsybl.loadflow.LoadFlowParameters;
 import com.powsybl.math.matrix.MatrixFactory;
 import com.powsybl.math.matrix.SparseMatrixFactory;
 import com.powsybl.openloadflow.OpenLoadFlowParameters;
+import com.powsybl.openloadflow.sensi.DcSensitivityAnalysis.AbstractDcSensitivityAnalysis;
+import com.powsybl.openloadflow.sensi.DcSensitivityAnalysis.DcFastContingencyAnalysis;
 import com.powsybl.sensitivity.*;
 import com.powsybl.tools.PowsyblCoreVersion;
 import org.jgrapht.alg.util.Pair;
@@ -33,7 +35,7 @@ public class OpenSensitivityAnalysisProvider implements SensitivityAnalysisProvi
 
     private static final String NAME = "OpenSensitivityAnalysis";
 
-    private final DcSensitivityAnalysis dcSensitivityAnalysis;
+    private final AbstractDcSensitivityAnalysis dcSensitivityAnalysis;
 
     private final AcSensitivityAnalysis acSensitivityAnalysis;
 
@@ -42,7 +44,7 @@ public class OpenSensitivityAnalysisProvider implements SensitivityAnalysisProvi
     }
 
     public OpenSensitivityAnalysisProvider(MatrixFactory matrixFactory) {
-        dcSensitivityAnalysis = new DcSensitivityAnalysis(matrixFactory);
+        dcSensitivityAnalysis = new DcFastContingencyAnalysis(matrixFactory);
         acSensitivityAnalysis = new AcSensitivityAnalysis(matrixFactory);
     }
